requirements.txt
Flask==3.0.3
tensorflow-cpu==2.12.0
numpy==1.26.4
Pillow==10.4.0
Werkzeug==3.0.3
gunicorn==22.0.0
scikit-learn==1.5.1
matplotlib==3.9.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
email-validator==2.1.0
seaborn==0.13.2
pandas==2.2.2
opencv-python==4.10.0.84
tqdm==4.66.4
albumentations==1.4.11
imgaug==0.4.0

